+++ adapters/mqtt-vertx-base/src/main/java/org/eclipse/hono/adapter/mqtt/MessageMapping.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -13,15 +13,17 @@
 
 package org.eclipse.hono.adapter.mqtt;
 
+import org.eclipse.hono.adapter.client.command.Command;
 import org.eclipse.hono.util.ExecutionContext;
 import org.eclipse.hono.util.RegistrationAssertion;
 import org.eclipse.hono.util.ResourceIdentifier;
 
 import io.vertx.core.Future;
+import io.vertx.core.buffer.Buffer;
 
 /**
- * A service for processing messages uploaded by devices before they are being
- * forwarded downstream.
+ * A service for either processing messages uploaded by devices before they are being
+ * forwarded downstream or mapping payload for commands to be sent to devices.
  *
  * @param <T> The type of execution context supported by this mapping service.
  */
@@ -47,8 +49,31 @@ public interface MessageMapping<T extends ExecutionContext> {
      *                         if the message could not be mapped.
      * @throws NullPointerException if any of the parameters are {@code null}.
      */
-    Future<MappedMessage> mapMessage(
+    Future<MappedMessage> mapDownstreamMessage(
             T ctx,
             ResourceIdentifier targetAddress,
             RegistrationAssertion registrationInfo);
+
+    /**
+     * Maps a command to be sent to a device/gateway.
+     * <p>
+     * If a command mapping service is not configured for a gateway or a protocol adapter, this method returns
+     * a successful future containing the original command payload. If a command mapping service is configured for a gateway
+     * or a protocol adapter and the mapping service returns a 200 OK HTTP status code, then this method returns a successful
+     * future containing the mapped command.
+     * <p>
+     * For all other 2XX HTTP status codes, this method returns the original command payload.
+     * In all other cases, this method returns a failed future with a {@link org.eclipse.hono.client.ServiceInvocationException}.
+     *
+     * @param registrationInfo The information included in the registration assertion for
+     *                         the gateway/device to which the command needs to be sent.
+     * @param command The original command to be mapped.
+     * @return                 A successful future containing the mapped command.
+     *                         Otherwise, the future will be failed with a {@link org.eclipse.hono.client.ServiceInvocationException}
+     *                         if the command could not be mapped.
+     * @throws NullPointerException if any of the parameters are {@code null}.
+     */
+    Future<Buffer> mapUpstreamMessage(
+        RegistrationAssertion registrationInfo,
+        Command command);
 }